Speedups for jtag read
Added buffering for FTDI I/O and refactored a bit. Reading is still slow
but I suspect this is down to the FTDI driver itself. Interestingly the
windows version is significantly faster than the linux one. The next step
is to group the reads into blocks rather than 32 bits at a time. Laters.